+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/onboarding/auth/data/error/FirebaseAuthErrorHandler.kt
@@ -1,0 +1,5 @@
+package com.poli.health.aquamate.onboarding.auth.data.error
+
+internal interface FirebaseAuthErrorHandler {
+    fun getErrorMessage(exception: Exception): String
+}