+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/onboarding/auth/di/IntakePresentationModule.kt
@@ -1,0 +1,18 @@
+package com.poli.health.aquamate.onboarding.auth.di
+
+import com.poli.health.aquamate.onboarding.auth.presentation.viewmodel.IntakeViewModel
+import org.koin.core.module.Module
+import org.koin.core.module.dsl.viewModel
+import org.koin.dsl.module
+
+val intakePresentationModule: Module = module {
+
+    viewModel {
+        IntakeViewModel(
+            registerWaterIntakeUseCase = get(),
+            getLastIntakeUseCase = get(),
+            deleteLastIntakeUseCase = get(),
+            getTodayTotalUseCase = get()
+        )
+    }
+}