+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/onboarding/auth/domain/usecase/DeleteLastIntakeUseCase.kt
@@ -1,0 +1,5 @@
+package com.poli.health.aquamate.onboarding.auth.domain.usecase
+
+interface DeleteLastIntakeUseCase {
+    suspend operator fun invoke(): Result<Unit>
+}