+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/onboarding/auth/domain/usecase/GetCurrentUserIdUseCase.kt
@@ -1,0 +1,8 @@
+package com.poli.health.aquamate.onboarding.auth.domain.usecase
+
+import kotlinx.coroutines.flow.Flow
+
+interface GetCurrentUserIdUseCase {
+
+    operator fun invoke(): Flow<String?>
+}