+++ composeApp/src/commonMain/kotlin/com/poli/health/aquamate/onboarding/auth/domain/usecase/SignOutUseCase.kt
@@ -1,0 +1,6 @@
+package com.poli.health.aquamate.onboarding.auth.domain.usecase
+
+interface SignOutUseCase {
+
+    suspend operator fun invoke()
+}