+++ composeApp/src/iosMain/kotlin/com/poli/health/aquamate/KoinInitializer.kt
@@ -1,0 +1,20 @@
+package com.poli.health.aquamate
+
+import com.poli.health.aquamate.framework.platformModule
+import com.poli.health.aquamate.onboarding.auth.di.authModule
+import com.poli.health.aquamate.onboarding.auth.di.intakeModule
+import org.koin.core.context.startKoin
+import org.koin.mp.KoinPlatformTools
+
+fun initKoin() {
+    if (KoinPlatformTools.defaultContext().getOrNull() == null) {
+        startKoin {
+            modules(
+                platformModule,
+                authModule,
+                intakeModule
+            )
+        }
+    }
+}
+