+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## Version: 0.0.1
+## Version: 1.0.0
 ### Added
 - The theme that this application will use has been implemented.
 - Implemented Koin to dependency injection and Firebase with auth library.
+- Implement auth module (sign in, sign up, sign out) in the app.
+- Implement Datastore (jetpack) to persist local data.
+- Implemented firebase from gitlive. This tool allows use firebase features.
+- Implemented screen of measure daily
+- Implemented navigation in auth y measure daily screens